Move the spaceship drawing code out of mapview_common
It doesn't have anything to do with the map.
