+++ client/mapview_common.cpp
@@ -2580,89 +2580,6 @@ void free_mapcanvas_and_overview()
   delete mapview.tmp_store;
 }
 
-/**
-   Return the desired width of the spaceship canvas.
- */
-void get_spaceship_dimensions(int *width, int *height)
-{
-  auto sprite = get_spaceship_sprite(tileset, SPACESHIP_HABITATION);
-  *width = 7 * sprite->width();
-  *height = 7 * sprite->width();
-}
-
-/**
-   Draw the spaceship onto the canvas.
- */
-void put_spaceship(QPixmap *pcanvas, int canvas_x, int canvas_y,
-                   const struct player *pplayer)
-{
-  Q_UNUSED(canvas_x)
-  Q_UNUSED(canvas_y)
-  int i, x, y;
-  const struct player_spaceship *ship = &pplayer->spaceship;
-  const QPixmap *spr;
-  struct tileset *t = tileset;
-
-  spr = get_spaceship_sprite(t, SPACESHIP_HABITATION);
-  const int w = spr->width(), h = spr->height();
-
-  QPainter p(pcanvas);
-  p.fillRect(0, 0, w * 7, h * 7,
-             get_color(tileset, COLOR_SPACESHIP_BACKGROUND));
-
-  for (i = 0; i < NUM_SS_MODULES; i++) {
-    const int j = i / 3;
-    const int k = i % 3;
-
-    if ((k == 0 && j >= ship->habitation)
-        || (k == 1 && j >= ship->life_support)
-        || (k == 2 && j >= ship->solar_panels)) {
-      continue;
-    }
-    x = modules_info[i].x * w / 4 - w / 2;
-    y = modules_info[i].y * h / 4 - h / 2;
-
-    spr = (k == 0   ? get_spaceship_sprite(t, SPACESHIP_HABITATION)
-           : k == 1 ? get_spaceship_sprite(t, SPACESHIP_LIFE_SUPPORT)
-                    : get_spaceship_sprite(t, SPACESHIP_SOLAR_PANEL));
-    p.drawPixmap(x, y, *spr);
-  }
-
-  for (i = 0; i < NUM_SS_COMPONENTS; i++) {
-    const int j = i / 2;
-    const int k = i % 2;
-
-    if ((k == 0 && j >= ship->fuel) || (k == 1 && j >= ship->propulsion)) {
-      continue;
-    }
-    x = components_info[i].x * w / 4 - w / 2;
-    y = components_info[i].y * h / 4 - h / 2;
-
-    spr = ((k == 0) ? get_spaceship_sprite(t, SPACESHIP_FUEL)
-                    : get_spaceship_sprite(t, SPACESHIP_PROPULSION));
-
-    p.drawPixmap(x, y, *spr);
-
-    if (k && ship->state == SSHIP_LAUNCHED) {
-      spr = get_spaceship_sprite(t, SPACESHIP_EXHAUST);
-      p.drawPixmap(x + w, y, *spr);
-    }
-  }
-
-  for (i = 0; i < NUM_SS_STRUCTURALS; i++) {
-    if (!BV_ISSET(ship->structure, i)) {
-      continue;
-    }
-    x = structurals_info[i].x * w / 4 - w / 2;
-    y = structurals_info[i].y * h / 4 - h / 2;
-
-    spr = get_spaceship_sprite(t, SPACESHIP_STRUCTURAL);
-    p.drawPixmap(x, y, *spr);
-  }
-
-  p.end();
-}
-
 /****************************************************************************
   Map link mark module: it makes link marks when a link is sent by chating,
   or restore a mark with clicking a link on the chatline.
